Reduce padding around text buttons
After moving the padding to the content of buttons, we ended up with
extra padding around text buttons that wasn't there before. The
"Download" bar doesn't look great with no padding at all, but it needs
to match the text buttons so the text doesn't jump to the right when
it switches from "Download" (text button) to "Connecting" (progress bar).

So we pad both text buttons and progress bars just a little bit, less
than a standard button but enough that progress bars look OK.