Don't show error message if price for a file can't be retrieved
On search results pages, stream_cost_estimate is called on every
result. If there's some problem in the daemon that affects all calls
to stream_cost_estimate, this can lead to 25 error boxes. Better to
just suppress the error and not display a price.

Also switched to the new name for the daemon method (get_cost_est ->
stream_cost_estimate)