Fix refreshing behavior on new input in WunderBar
Before, <WunderBar /> would compare the address inside it to the page
address passed in by <Header /> to determine if there was a page change.
This would cause user input to be replaced with "Home" when typing a
search, since during a search the page address is always "Home," not
the query.

Now it simply checks if the page being viewed has changed.