Update view.jsx
Small idea how  "Help" section could be improved with a link to status.lbry.io. It could help the users to identify any current slow-down or problems. I was not sure what would be the appropriate button icon so I just left the "question mark" one. Feel free to dismiss this idea as not important or change any wording.