WIP: wallet encrypt and decrypt with sync There's a flow error you can hopefully fix up.
There seems to be an odd case when you have a synced wallet, then turn off sync, and then change the password. When you go back to turn it on, it will prompt for the old password and everything gets set back there. I'm thinking we ask to enable sync on the wallet encrypt page also.

Conflicts:
	src/ui/modal/modalWalletEncrypt/index.js
	src/ui/modal/modalWalletEncrypt/view.jsx
	src/ui/page/settings/view.jsx