Changing some text on the publishing page
Changed "Your video may not be the best format. Use MP4s in H264/AAC format and a friendly bitrate (720p) for more reliable streaming." to "Your video may not be the best format. Use MP4s in H264/AAC format and a friendly bitrate (under 5 mbps) and resolution (720p) for more reliable streaming." and changed the two times it says this "For video content, use MP4s in H264/AAC format and a friendly bitrate (720p) for more reliable streaming." to "For video content, use MP4s in H264/AAC format and a friendly bitrate (under 5 mbps) and resolution (720p) for more reliable streaming."