Closes issue #4192
This si second atempt, because we must use different names for the text for "Following" from the channel page and another text for the sidebar panel and another for searh list options. Currently is the same text everywhere and can't be translated separate.
*And most interestingly*:
- the message that a new user is receiving after first channel **subscription** is "Awesome! You just subscribed to your first channel." with the label "Subscriptions 101" as you can see in the file ui\modal\modalFirstSubscription\viewjs.
- everywhere in the app (in over 20 files) is used the word subscriptions and subscribed as in:
 (import { selectSubscriptions } from 'redux/selectors/subscriptions';)
(subscriptions: selectSubscriptions(state),)
(subscriptions: subscriptionsReducer,)
- many commented lines in the code like:
     //   Populate a channels unread subscriptions or update the type
     // Remove multiple files (or all) from a channels unread subscriptions
     // If no channel is passed in, remove all unread subscriptions from all channels
    // if the user isn't sharing data, keep the subscriptions entirely in the app
   // They are sharing data, we can store their subscriptions in our internal database
    Lbryio.call('subscription', 'new', {
      channel_name: subscription.channelName, etc.
- even the files from redux (actions, reducers, selectors) folder are named subscriptions.js
- ui\store.js uses: const subscriptionsFilter = createFilter('subscriptions', ['subscriptions']); and const whiteListedReducers = [...   'subscriptions',];
and most interesting:
const sharedStateFilters = {
  tags: { source: 'tags', property: 'followedTags' },
  subscriptions: {
    source: 'subscriptions',
    property: 'subscriptions',
    transform: function(value) {
      return value.map(({ uri }) => uri);
    },
where we can see the use of the "subscriptions" and "followed tags" string also.

So with so many examples in the code, why not use the same term in the program interface too? But the main reason is that we need to separate the instances where the "following" word is used.

An alternative for using "Subscriptions" is that we can use "Followed channels", but I think is too long. So do you have any real reason for not using Subscriptions? After all, we must use other word in side panel instead of the "Following" which is inappropriate at least in other languages.