Form-field: Truncate the "read-only prefix" to 40% width so that the input field is always visible.
Currently, the input field goes off-screen if the read-only prefix is too long.

Given that the prefix is read-only, the user probably knows what the prefix is from context. e.g. In the case of the Repost Modal, the prefix is repeated in a selection field above. So, it's probably ok to just truncate the text to an ellipsis.

The input field is the more important one here, so giving it 60% width when truncation is needed.