Limit the Markdown Editor width in Edit (not New) Mode to prevent overflow.
The MDE used in the "new" section has a parent width parameter to limit itself, while the MDE used in the "edit" section didn't.

Fix by limiting ".comment__body_container" to 80%, which takes into account the space taken by the author's avatar. This feels a little bit dirty since it's hard-coded.  If there's a way to calculate the avatar width from here, it will be more robust.