Translate the footer.
Most of the items in the footer don't need to be translated, but it was troublesome trying to make `sections[]` re-resolve the label on a language change, if we want to selectively translate certain labels only.

So, just translate everything for now. Translators can simply clone the English string if it doesn't need a translation.