Fix sidebar tooltip showing string with context metadata.
## Issue
In the odysee branch, the tooltip for "Following" in the Sidebar was shown as "Following --[sidebar]--". In lbry.tv, the tooltip doesn't appear.

I believe the tooltip appeared due to the name change from 'label' to 'title' and then being passed to Button through '...passedProps'.

Fix by explicitly setting the Button's 'title' to the localized text.