vjs i18n: Fix "undefined" error due to hidden ctrls
## Issue
The change in making the "Quality" component hidden by default exposed a minor error in `resolveCtrlText` for not handling hidden components.

Luckily, videojs cathes the error, so it didn't crash the app.

## Change
Add a try-catch block to handle any temp-removal of components in the future.

## Note
This is a touch up for PR_5584