Remove the need to call 'app.getLocale()'
The removal is primarily to ease the upcoming language refactoring, but it was also never hit anyway, because if the `storedLanguage` is indeed null or undefined (like in the case of a fresh install), the `then` block never runs.

Fortunately, we do still get the effect of "falling back to OS's language", because `selectLanguage` will call `getDefaultLanguage` if `language` is null. While the `language` localStorage remains null, the rest of the code still gets the native language via `selectLanguage`.

Aside: `getDefaultLanguage` uses `navigator.language`. It does seem to grab my OS's locale correctly.